Major emmeans fix
Implement functionality for dropping experimental groups that are
unsupported by any samples. This necessitated wrapping the emmeans
operations so that it had access to the full grid (which isn't
possible when the design is pre-expanded to a matrix).

Also refactored specification file so that lists of things are plural,
and individual items have their own alias for 'list' to make it more
intuitive.