Move param settings into the spec file
The 'settings' section of the spec file will contain the analysis
parameters that fine-tune the analysis (alpha's, etc).  These are now
fed into the param object as a 'default' - it remains necessary to set
them before they are used, but if the value is missing the value from
the spec file will be picked up.