Fix #5 unused differential_heatmap arguments
I've removed the tidy_fn from the 01_analyse.r template. As it
defaults to NULL in the differential_heatmap function, and doesn't get
used there, this should not break anything or change results. The
behaviour was initially there to allow a user-specified
white-centring, but the automatic centring has worked ok for a while
so this argument was orphaned.  I'll leave the functionality in, as
eventually it will be desirable to have a user-specifiable option
here.  The template called differential_heatmap with a positional
argument for tidy_fn, so if/when we re-activate it, probably best to
re-order the arguments and put a check on the second one `match.call`
should handle this