Major changes to allow direct generation of the report on a dds object
In addition to the extra `run_report` command, there is now better
handling of missing mcols or metadata, so things don't unnecessarily
break if they're not provided.

`spec_suffix` has now been renamed as `count_source` - this is a more
meaningful name but might need some retro-fitting if people upgrade to
this new DESdemonA without using the new makefiles.